logrittr logger for lumberjack

An R6 logger that plugs into the lumberjack %L>% pipe and renders the same console output as %>=%: row/column counts, signed deltas, added/dropped column names, and approximate step timing.

Requires the R6 and lumberjack packages (both in Suggests). If either is missing, the class is unavailable but the rest of logrittr works normally.

Timing

lumberjack calls $add() after each step without providing a start time. Elapsed time is measured as the interval between two consecutive $add() calls. The first step always shows NA ms.

See also

Methods

Public methods

logrittr_logger$new()

logrittr_logger$add()

logrittr_logger$dump()

logrittr_logger$clone()

Method new()

Create a new logrittr_logger.

Usage

logrittr_logger$new()

Method add()

Called by lumberjack after each pipe step.

Usage

logrittr_logger$add(meta, input, output)

Arguments

meta

List with elements expr and src (the step expression).

input

Data frame before the step.

output

Data frame after the step.

Method dump()

Called by dump_log(). No-op: output is already streamed to the console in real time.

Usage

logrittr_logger$dump(...)

Arguments

...

Ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

logrittr_logger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
library(lumberjack)
library(dplyr)

iris %L>%
  start_log(log = logrittr_logger$new()) %L>%
  as_tibble() %L>%
  filter(Sepal.Length < 5) %L>%
  mutate(rn = row_number()) %L>%
  group_by(Species) %L>%
  summarise(n = n_distinct(rn)) %L>%
  dump_log(stop = TRUE)
  
  
# logfile <- tempfile(fileext=".r.log.txt")
# 
#  iris %L>%
#   start_log(log = logrittr_logger$new()) %L>%
#   as_tibble()  %L>%
#   filter(Sepal.Length < 5) %L>%
#   mutate(rn = row_number()) %L>%
#   group_by(Species) %L>%
#   summarise(n = n_distinct(rn)) %L>%
#   dump_log(file=logfile)
#   
#   logdata <- read.lines(logfile)
#   
#   head(logdata)
  
} # }